portfolio/demo-rag-chatbot/requirements.txt
fastapi
uvicorn[standard]
python-multipart
pypdf
langchain
langchain-openai
langchain-chroma
langchain-text-splitters
chromadb
python-dotenv
